whew mostly minor bug fixes, but likely some bugs too ;)

+++ inc/Player.h
@@ -1,6 +1,8 @@
 #ifndef PLAYER_H_
 #define PLAYER_H_
 
+#include <cstdint>
+
 #include "Card.h"
 
 /**
@@ -14,9 +16,9 @@ public:
   Player(std::string name = "") : name_(name) { }
 
   inline std::string getName() const { return name_; }
-  inline size_t getChips() const { return chips_; }
-  inline size_t getChipsInPlay() const { return chips_in_play_; }
-  inline size_t getSeat() const { return seat_; }
+  inline uint32_t getChips() const { return chips_; }
+  inline uint32_t getChipsInPlay() const { return chips_in_play_; }
+  inline uint32_t getSeat() const { return seat_; }
   
   // for this to be correct, rely on Game invariant of removing Players
   // that have lost all their chips as soon as possible.
